Cleanup code and consolidate Clock interface
* Replace Supplier<List> with just List when constructing
  RequestDispatcher.
* Remove always true Predicate in RetryingRequestDispatcher
* Make a defensive copy of Endpoints in EndpointIterator
* Remove an unused method
